% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seamos.R
\name{seamos}
\alias{seamos}
\title{Standardised Effects Adjusted for Multiple Overlapping Subgroups (SEAMOS)}
\usage{
seamos(data, outcome, treat, covs, nperm)
}
\arguments{
\item{data}{A data frame containing the trial data.}

\item{outcome}{A string indicating the name of the outcome variable.}

\item{treat}{A string indicating the randomised treatment variable. This should be coded as a numeric 0/1.}

\item{covs}{A vector of strings containing the names of the baseline covariates which define the subgroups.
So far these must all be binary 0/1 covariates.}

\item{nperm}{The number of permutations you want to use.}
}
\value{
A list containing the largest absolute effect size from the observed data and an array
containing the permutation distribution of the largest absolute effect sizes.
}
\description{
Standardised Effects Adjusted for Multiple Overlapping Subgroups (SEAMOS)
}
